coding guidelines: partially comply with MISRA C:2012 essential types rules.
In particular:

- use bool when the data nature is Boolean;

- avoid mixing signed and unsigned integers in computations and
  comparisons;

- avoid mixing enumerations with integers: when this is unavoidable,
  always convert enums to integers and not the other way around;

- avoid mixing characters with integers;

- ensure computations are done in the destination precision;

- added /*? comments when the developer intentions are not clear;

- added U suffix to unsigned constants (except for the CONFIG_* macro
  constants, as they cannot be changed and then their use as unsigned
  constants should be prefixed with a cast).

Violations for rules 10.1, 10.2, 10.3, 10.4, 10.5, 10.6, 10.7, 10.8,
11.7, 12.2, 14.4 and 16.7 in the reference builds have been reduced
from 67818 to 83.  The commit cannot be divided on a per-rule basis
due to numerous cross-dependencies between changes.

+++ include/arch/common/ffs.h
@@ -30,11 +30,11 @@ extern "C" {
 
 static ALWAYS_INLINE unsigned int find_msb_set(uint32_t op)
 {
-	if (op == 0) {
-		return 0;
+	if (op == 0U) {
+		return 0U;
 	}
 
-	return 32 - __builtin_clz(op);
+	return 32U - (unsigned int)__builtin_clz(op);
 }
 
 
@@ -53,7 +53,7 @@ static ALWAYS_INLINE unsigned int find_msb_set(uint32_t op)
 static ALWAYS_INLINE unsigned int find_lsb_set(uint32_t op)
 {
 #ifdef CONFIG_TOOLCHAIN_HAS_BUILTIN_FFS
-	return __builtin_ffs(op);
+	return (unsigned int)__builtin_ffs((int)op);
 
 #else
 	/*
